+++ oc_common.h
@@ -6,11 +6,11 @@
  * proto type of internal apis.
  *
  * @author	JoungKyun.Kim <http://oops.org>
- * $Date: 2011-03-29 10:08:14 $
- * $Revision: 1.21.2.2 $
+ * $Date: 2011-03-29 18:18:35 $
+ * $Revision: 1.21.2.3 $
  * @attention	Copyright (c) 2011 JoungKyun.Kim all rights reserved.
  */
-/* $Id: oc_common.h,v 1.21.2.2 2011-03-29 10:08:14 oops Exp $ */
+/* $Id: oc_common.h,v 1.21.2.3 2011-03-29 18:18:35 oops Exp $ */
 
 #ifndef OC_COMMON_H
 #define OC_COMMON_H
@@ -81,7 +81,7 @@
 	#define ULLONG_MAX   18446744073709551615ULL
 #endif
 
-UInt get_charcount (CChar * str, CChar * del);
+size_t get_charcount (CChar * str, CChar * del);
 Bit64 devided64_high_low (ULong64 v);
 ULong64 combined64_high_low (Bit64 v);
